re-organize closet & outfit related pages
